python/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
geoalchemy2==0.17.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
redis==5.2.1
boto3==1.36.6
geopandas==1.0.1
shapely==2.0.6
h3==3.7.7
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.1
xgboost==2.1.3
shap==0.46.0
weasyprint==63.1
jinja2==3.1.5
grpcio==1.69.0
grpcio-tools==1.69.0
protobuf==5.29.3
python-multipart==0.0.20
geojson==3.2.0
pyproj==3.7.0
requests==2.32.3
beautifulsoup4==4.12.3
unidecode==1.3.8
alembic==1.14.1
rasterio==1.3.11
openpyxl==3.1.5
mcp[cli]>=1.2.0
